lift updateConflictState out to own module
